code cleanup moving js fingerprint tests, utilities and core to separate files
